fix(opnsense): widen byte/packet counter fields to int64 for 32-bit source builds
utils.go's string parsers were widened to int64 for 32-bit safety (commit
7f58a29, upstream #81), but the JSON-tagged struct fields were missed. On a
32-bit build (GOARCH=386/arm) Go's int is 32-bit, and real firewall byte
counters routinely exceed ~2.1B (live box: in4_pass_bytes=64.6G). Direct
JSON unmarshal of such a value into a plain int field returns an
UnmarshalTypeError that fails the whole subsystem fetch (firewall pf-stats,
firewall rules, ipsec phase1), while protocol_statistics' numToInt narrowed
via int(i), silently wrapping large TCP/UDP byte counters into negative
garbage.

Widen the byte/packet counter fields in firewall.go, firewall_rules.go,
ipsec.go (phase1 and phase2), mbuf.go, and the ProtocolStatistics struct
(incl. its map[string]int → map[string]int64) to int64, and make numToInt
return int64. ipsec phase2 byte/packet parsing switches to the tolerant
safeAtoi (int64) so large values are preserved rather than zeroed. Values
are converted to float64 at the Prometheus boundary, so metric output is
unchanged. Verified linux/386 and linux/arm cross-builds compile.

Closes #103

+++ internal/collector/firewall.go
@@ -192,7 +192,7 @@ func (c *firewallCollector) Update(ctx context.Context, client *opnsense.Client,
 	}
 
 	for _, v := range data.Interfaces {
-		metricsValueMapping := map[*prometheus.Desc]int{
+		metricsValueMapping := map[*prometheus.Desc]int64{
 			c.inIPv4PassPackets:   v.In4PassPackets,
 			c.outIPv4PassPackets:  v.Out4PassPackets,
 			c.inIPv4BlockPackets:  v.In4BlockPackets,
